chore: Refactor chat bubbles
