+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AgentChatAssistantBubble.tsx
@@ -2,12 +2,9 @@ import {
     AiResultType,
     parseVizConfig,
     type AiAgentMessageAssistant,
-    type AiAgentMessageUser,
-    type AiAgentUser,
 } from '@lightdash/common';
 import {
     ActionIcon,
-    Card,
     Center,
     CopyButton,
     Group,
@@ -27,13 +24,10 @@ import {
     IconThumbUpFilled,
 } from '@tabler/icons-react';
 import MDEditor from '@uiw/react-md-editor';
-import { format, parseISO } from 'date-fns';
 import { memo, useCallback, useMemo, type FC } from 'react';
 import { useParams } from 'react-router';
 import MantineIcon from '../../../../../components/common/MantineIcon';
 import { useInfiniteQueryResults } from '../../../../../hooks/useQueryResults';
-import { useTimeAgo } from '../../../../../hooks/useTimeAgo';
-import useApp from '../../../../../providers/App/useApp';
 import {
     useAiAgentThreadMessageVizQuery,
     useUpdatePromptFeedbackMutation,
@@ -45,53 +39,6 @@ import {
 import { isOptimisticMessageStub } from '../../utils/thinkingMessageStub';
 import { AiChartVisualization } from './AiChartVisualization';
 import { AiChartToolCalls } from './ToolCalls/AiChartToolCalls';
-
-export const UserBubble: FC<{ message: AiAgentMessageUser<AiAgentUser> }> = ({
-    message,
-}) => {
-    const timeAgo = useTimeAgo(message.createdAt);
-    const name = message.user.name;
-    const app = useApp();
-    const showUserName = app.user?.data?.userUuid !== message.user.uuid;
-
-    return (
-        <Stack gap="xs" style={{ alignSelf: 'flex-end' }}>
-            <Stack gap={0} align="flex-end">
-                {showUserName ? (
-                    <Text size="sm" c="gray.7" fw={600}>
-                        {name}
-                    </Text>
-                ) : null}
-                <Tooltip
-                    label={format(parseISO(message.createdAt), 'PPpp')}
-                    withinPortal
-                >
-                    <Text size="xs" c="dimmed">
-                        {timeAgo}
-                    </Text>
-                </Tooltip>
-            </Stack>
-
-            <Card
-                pos="relative"
-                radius="md"
-                py="xs"
-                px="sm"
-                withBorder={true}
-                bg="white"
-                color="white"
-                style={{
-                    overflow: 'unset',
-                }}
-            >
-                <MDEditor.Markdown
-                    source={message.message}
-                    style={{ backgroundColor: 'transparent' }}
-                />
-            </Card>
-        </Stack>
-    );
-};
 
 const AssistantBubbleContent: FC<{ message: AiAgentMessageAssistant }> = ({
     message,
