+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AgentChatDisplay.tsx
@@ -2,7 +2,8 @@ import { type AiAgentMessage, type AiAgentThread } from '@lightdash/common';
 import { Box, Divider, Flex, getDefaultZIndex, Stack } from '@mantine-8/core';
 import { Fragment, useRef, type FC, type PropsWithChildren } from 'react';
 import ErrorBoundary from '../../../../../features/errorBoundary/ErrorBoundary';
-import { AssistantBubble, UserBubble } from './AgentChatBubbles';
+import { AssistantBubble } from './AgentChatAssistantBubble';
+import { UserBubble } from './AgentChatUserBubble';
 import ThreadScrollToBottom from './ScrollToBottom';
 import { ChatElementsUtils } from './utils';
